requirements.txt
discord.py>=2.5.0
python-dotenv>=1.0.0
pymongo[srv]>=4.0.0
dnspython>=2.3.0
certifi>=2023.0.0
Pillow>=10.0.0
